Restructuring to reduce the complexity of the Entry Point
