+++ agents/base/Agent.java
@@ -2,7 +2,7 @@ package grakn.simulation.agents.base;
 
 import grabl.tracing.client.GrablTracingThreadStatic.ThreadContext;
 import grakn.simulation.agents.RandomValueGenerator;
-import grakn.simulation.agents.World;
+import grakn.simulation.world.World;
 import grakn.simulation.driver.DbDriverWrapper;
 import graql.lang.query.GraqlQuery;
 import org.slf4j.Logger;
