+++ agents/common/CountryAgent.java
@@ -1,8 +1,7 @@
 package grakn.simulation.agents.common;
 
-import grakn.simulation.agents.World;
+import grakn.simulation.world.World;
 import grakn.simulation.agents.base.Agent;
-import grakn.simulation.agents.base.AgentRunner;
 
 public abstract class CountryAgent extends Agent<World.Country> {
