+++ world/BUILD
@@ -1,0 +1,10 @@
+java_library(
+    name = "world",
+    srcs = glob([
+        "*.java",
+    ]),
+    visibility = ["//visibility:public"],
+    deps = [
+        "@maven//:org_apache_commons_commons_csv",
+    ],
+)
